refactor: create providers configuration to filter enabled agents for BalanceProvider
